Refactoring of Font system
Font is now a simple list of Glyph, and can return the glyph corresponding to a give char.

Font6 and Font9 do not inherited of Font anymore (which was odd from the start). Instead, they build a font and offer a method to get it.

Maybe this method should be an interface. I'm not sure it really matter yet.